fix bugs. some updates.

+++ HPS/printer/Controller/GCodeCommands.cpp
@@ -17,7 +17,7 @@ void PrinterController::calc_steps_speed(float dx, float dy, float dz, float de,
     //позволяет настроить скорость и время для единовременного завершения работы двигателей
     float max = abs(steps_z);
     float speed = fmin(maxspeed, position.s);
-    int32_t n_speed = int(speed/60);
+    float n_speed = speed / 60;
     float t = abs(dz/n_speed); //t - общее время при макс скорости в секундах
     if (abs(steps_a) > max)
     {   max = abs(steps_a);
@@ -220,7 +220,6 @@ void PrinterController::gcode_M140(const Parameters& parameters) {
 }
 
 void PrinterController::gcode_M190(const Parameters& parameters) {
-    //температура необязательная. Исправить
     if (parameters.find('S'))
         mechanics.wait_bed_temperature(parameters['S']);
     else
